pyproject.toml
[project]
name = "mcp-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "mcp[cli]==1.10.1",
    "nest-asyncio>=1.6.0",
    "openai==1.75.0",
    "python-dotenv>=1.2.2",
    "streamlit>=1.56.0",
    "youtube-transcript-api>=1.2.4",
    "yt-dlp>=2026.3.17",
]

[tool.pylance]
extraPaths = ["servers/youtube"]
